refactor: re-org VaultStore modules
