+++ src/VaultStore/helpers/database.ts
